Moved datadir locking up from client to node
Moves the logic for acquiring a lock on datadir to happen
before client authentication. This allows for earlier detection of a Golem instance
running on the machine.
